migrate bone
